pyproject.toml
[tool.poetry]
name = "bizcate-model"
version = "0.1.0"
description = ""
authors = ["fuser-vijay <vijaykumar.m@fusionanalytics.com>"]
readme = "README.md"
packages = [{include = "bizcate_model"}]

[tool.poetry.dependencies]
python = ">3.8,<4.0"
siuba = "^0.4.2"
python-dotenv = "^1.0.0"
fusion-kf = {git = "https://github.com/Fusion-Tools/fusion_kf"}
fusion-db = {git = "https://github.com/Fusion-Tools/fusion_db.git"}

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
